requirements.txt
uvicorn
fastapi[standard]
typing
pedantic
pandas
mysql-connector-python
# pymongo
# dnspython
# mysqlclient
